refactor: :truck: rename files to `extract_field_properties()`
