+++ src/seedcase_sprout/extract_field_properties.py
